# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Event Management API
  version: 1.0.0
  description: API documentation for the Event Management System
servers:
  - url: http://localhost:3000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        firstname:
          type: string
        lastname:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [CUSTOMER, ORGANIZER, ADMIN]
        phone:
          type: string
        createdAt:
          type: string
          format: date-time
    Profile:
      type: object
      properties:
        profileId:
          type: string
        userId:
          type: string
        firstname:
          type: string
        lastname:
          type: string
        phone:
          type: string
        city:
          type: string
        subcity:
          type: string
        housenumber:
          type: string
        profilePicture:
          type: string
          description: URL or path to the profile picture
        createdAt:
          type: string
          format: date-time
    Category:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
    events:
      type: object
      properties:
        eventId:
          type: string
        title:
          type: string
        description:
          type: string
        date:
          type: string
          format: date-time
        time:
          type: string  
        location:
          type: string
        capacity:
          type: integer
        status:
          type: string
          enum: [UPCOMING, ONGOING, COMPLETED, CANCELLED]
        organizerId:
          type: string
        categoryId:
          type: string
        createdAt:
          type: string
          format: date-time
    registrations:
      type: object
      properties:
        userId:
          type: string
        eventId:
          type: string
        registration_date:
          type: string
          format: date-time
        status:
          type: string
          enum: [PENDING, CONFIRMED, CANCELLED]
        cancelled_at:
          type: string
          format: date-time
    notifications:
      type: object
      properties:
        notificationId:
          type: string
        userId:
          type: string
        event_id:
          type: string
        message:
          type: string
        type:
          type: string
          enum: [REMAINDER, UPDATE, CAMCELLATION]  
        is_read:
          type: boolean
        sent_at:
          type: string
          format: date-time
...
    feedback:
      type: object
      properties:
        feedback_id:
          type: string
        user_id:
          type: string
        event_id:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
        submitted_at:
          type: string
          format: date-time    
    payments:
      type: object
      properties:
        payment_id:
          type: string
        user_id:
          type: string
        event_id:
          type: string
        amount:
          type: number
          format: float
        payment_method:
          type: string
          enum: [CREDIT_CARD, PAYPAL, BANK_TRANSFER]
        transaction_reference:
          type: string  
        payment_status:
          type: string
          enum: [PENDING, SUCCESS, FAILED]
        paid_at:
          type: string
          format: date-time        
paths:
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - firstname
                - lastname
                - email
                - password
                - role
                - phone
              properties:
                firstname:
                  type: string
                  example: "Yeamlak"
                lastname:
                  type: string
                  example: "sisay"
                email:
                  type: string
                  example: "yeamlak.sisay@example.com"
                  format: email
                password:
                  type: string
                  minLength: 8
                role:
                  type: string
                  enum: [CUSTOMER, ORGANIZER, ADMIN]
                  description: Defaults to CUSTOMER if not provided
                phone:
                  type: string
                  example: "+251912345678"
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
                      token:
                        type: string
        '400':
          description: Missing/invalid fields
        '409':
          description: Email already exists
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Log in a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
                      token:
                        type: string
        '400':
          description: Missing email/password
        '401':
          description: Invalid email or password
  /api/auth/logout:
    post:
      tags: [Auth]
      summary: Log out a user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logout successful
        '401':
          description: Missing, invalid, or already-revoked token
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Get the current authenticated user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
        '401':
          description: Missing or invalid token
        '404':
          description: User not founds
  /api/users:
    get:
      tags: [Users]
      summary: List all users
      description: Admin only. Non-admin users will receive a 403.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      users:
                        type: array
                        items:
                          $ref: '#/components/schemas/User'
        '401':
          description: Missing or invalid token
        '403':
          description: Not an admin
...
  /api/categories:
    get:
      tags: [Categories]
      summary: Get all event categories
      description: Public endpoint, no auth required.
      responses:
        '200':
          description: List of categories
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      categories:
                        type: array
                        items:
                          $ref: '#/components/schemas/Category'

    post:
      tags: [Categories]
      summary: Create a new category
      description: Admin only.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        '201':
          description: Category created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      category:
                        $ref: '#/components/schemas/Category'
        '400':
          description: Missing/invalid fields
        '401':
          description: Missing or invalid token
        '403':
          description: Not an admin
        '409':
          description: Category already exists
...
  /api/profiles:
    post:
      tags: [Profiles]
      summary: Create the profile for a user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - firstname
                - lastname
                - phone
                - city
              properties:
                firstname:
                  type: string
                lastname:
                  type: string
                phone:
                  type: string
                city:
                  type: string
                subcity:
                  type: string
                housenumber:
                  type: string
                profilePicture:
                  type: string
                  description: URL or path to the profile picture
      responses:
        '201':
          description: Profile created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      profile:
                        $ref: '#/components/schemas/Profile'
        '400':
          description: Missing/invalid fields
        '401':
          description: Missing or invalid token
        '409':
          description: Profile already exists for this user
...
  /api/events:
    get:
      tags: [Events]
      summary: List all events
      description: Public endpoint, no auth required. Supports pagination and filtering.
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
          description: Page number
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
          description: Number of results per page
        - in: query
          name: categoryId
          schema:
            type: string
          description: Filter by category ID
        - in: query
          name: status
          schema:
            type: string
            enum: [UPCOMING, ONGOING, COMPLETED, CANCELLED]
          description: Filter by event status
        - in: query
          name: search
          schema:
            type: string
          description: Search by event title
      responses:
        '200':
          description: List of events
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      events:
                        type: array
                        items:
                          $ref: '#/components/schemas/events'
                      pagination:
                        type: object
                        properties:
                          total:
                            type: integer
                          page:
                            type: integer
                          limit:
                            type: integer
                          totalPages:
                            type: integer

    post:
      tags: [Events]
      summary: Create a new event
      description: Organizer only.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - date
                - time
                - location
                - capacity
                - categoryId
              properties:
                title:
                  type: string
                description:
                  type: string
                date:
                  type: string
                  format: date-time
                time:
                  type: string
                location:
                  type: string
                capacity:
                  type: integer
                categoryId:
                  type: string
      responses:
        '201':
          description: Event created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      event:
                        $ref: '#/components/schemas/events'
        '400':
          description: Missing/invalid fields
        '401':
          description: Missing or invalid token
        '403':
          description: Not an organizer
...
  /api/feedback:
    post:
      tags: [Feedback]
      summary: Submit feedback for an event
      description: Authenticated users only.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - event_id
                - rating
              properties:
                event_id:
                  type: string
                rating:
                  type: integer
                  minimum: 1
                  maximum: 5
                comment:
                  type: string
      responses:
        '201':
          description: Feedback submitted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      feedback:
                        $ref: '#/components/schemas/feedback'
        '400':
          description: Missing/invalid fields
        '401':
          description: Missing or invalid token
        '404':
          description: Event not found
        '409':
          description: Feedback already submitted for this event
...
  /api/notifications:
    get:
      tags: [Notifications]
      summary: List notifications for the current user
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
          description: Page number
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
          description: Number of results per page
        - in: query
          name: read
          schema:
            type: boolean
          description: Filter by read/unread status
      responses:
        '200':
          description: List of notifications
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      notifications:
                        type: array
                        items:
                          $ref: '#/components/schemas/notifications'
                      pagination:
                        type: object
                        properties:
                          total:
                            type: integer
                          page:
                            type: integer
                          limit:
                            type: integer
                          totalPages:
                            type: integer
        '401':
          description: Missing or invalid token

    post:
      tags: [Notifications]
      summary: Create a notification
      description: Admin only.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user_id
                - message
              properties:
                user_id:
                  type: string
                message:
                  type: string
                type:
                  type: string
      responses:
        '201':
          description: Notification created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      notification:
                        $ref: '#/components/schemas/notifications'
        '400':
          description: Missing/invalid fields
        '401':
          description: Missing or invalid token
        '403':
          description: Not an admin
        '404':
          description: Target user not found
...
  /api/payments:
    get:
      tags: [Payments]
      summary: List all payments
      description: Admin only.
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
          description: Page number
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
          description: Number of results per page
        - in: query
          name: payment_status
          schema:
            type: string
            enum: [PENDING, SUCCESS, FAILED]
          description: Filter by payment status
      responses:
        '200':
          description: List of payments
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      payments:
                        type: array
                        items:
                          $ref: '#/components/schemas/payments'
                      pagination:
                        type: object
                        properties:
                          total:
                            type: integer
                          page:
                            type: integer
                          limit:
                            type: integer
                          totalPages:
                            type: integer
        '401':
          description: Missing or invalid token
        '403':
          description: Not an admin

    post:
      tags: [Payments]
      summary: Create a payment
      description: Authenticated users only.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - event_id
                - amount
                - payment_method
              properties:
                event_id:
                  type: string
                amount:
                  type: number
                  format: float
                payment_method:
                  type: string
                  enum: [CREDIT_CARD, PAYPAL, BANK_TRANSFER]
      responses:
        '201':
          description: Payment created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      payment:
                        $ref: '#/components/schemas/payments'
        '400':
          description: Missing/invalid fields
        '401':
          description: Missing or invalid token
        '404':
          description: Event not found
...
  /api/registrations:
    get:
      tags: [Registrations]
      summary: List all registrations
      description: Admin only.
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
          description: Page number
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
          description: Number of results per page
        - in: query
          name: status
          schema:
            type: string
            enum: [PENDING, CONFIRMED, CANCELLED]
          description: Filter by registration status
      responses:
        '200':
          description: List of registrations
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      registrations:
                        type: array
                        items:
                          $ref: '#/components/schemas/registrations'
                      pagination:
                        type: object
                        properties:
                          total:
                            type: integer
                          page:
                            type: integer
                          limit:
                            type: integer
                          totalPages:
                            type: integer
        '401':
          description: Missing or invalid token
        '403':
          description: Not an admin

    post:
      tags: [Registrations]
      summary: Register for an event
      description: Authenticated users only.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - eventId
              properties:
                eventId:
                  type: string
      responses:
        '201':
          description: Registration created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      registration:
                        $ref: '#/components/schemas/registrations'
        '400':
          description: Missing/invalid fields
        '401':
          description: Missing or invalid token
        '404':
          description: Event not found
        '409':
          description: Already registered for this event, or event at full capacity
...